Fix project dev setup for MacOS
MacOS comes usually with old bash and sh versions, so some commands are
not exactly as in Linux.

Namely for our case, there's no `-t` option for `mv`, and the `-i` flag
of `find` requires always a value after it.

The simplest solution is to just have 2 different yarn scripts.