Hide crosstable and move-list in zen mode (#574)
They are distractions as well, so we should hide them too, same as
lichess does.

We hide each "move" instead of the whole #movelist to hide all the moves
but not the container that includes, for example the result of the game.
We don't hide the rematch/analysis buttons either. Again, same as lichess.