A hot end to the summer
![Some Birds](https://github.com/gbtami/pychess-variants/blob/master/static/images/AngryBirds.png)

We got a bunch of updates coming in to finish off this summer!

*New Variant*

We’re excited to announce a new variant at Pychess! This one involves taking your supply of birds (of varying colors and sizes) and throwing them at your opponent at various angles to bring their castle down as quickly and efficiently as possible! That’s right, we’re obviously talking about… Tori Shogi!

![Tori Shogi](https://github.com/gbtami/pychess-variants/blob/master/static/images/ShogiGuide/ToriIntl.png)
Tori Shogi is a three-centuries old variant that has similar to rules of Shogi but is played on a smaller board (7x7) and starts with immediate confrontation, also unlike Shogi. We’re now bringing it to the mainstream with piece symbols that should be more accessible to everyone. Please be sure to read [the rules](https://www.pychess.org/variants/torishogi) before you play because all the pieces are new and somewhat complex. It will be *much* faster to read about the game than *winging it*.

Also, before anyone asks, no we will not add Atomic Tori Shogi. We might be in trouble with PETA if we had a game featuring exploding birds! On a similar note, there will no more planned for variants for a good while so that we can spend more attention on improving the site’s features. With nearly 40 variants available, we hope that’s enough to keep most people satisfied with a variety of selection!
<iframe width="560" height="315" src="https://www.youtube.com/embed/2WNrx2jq184" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>

*New Engine*

Fairy is upgraded now, to version 14, or at least a version very close to it depending on how testing pans out. Good luck on challenging even tougher levels of play!

*New Piece Sets*

For some older news we haven’t gotten to, CBurnett-style pieces are now available for every chess variant (including all the variants with new armies). CBurnett is the default style on Lichess, while our default is Merida. Feel free to play around the settings and try out some variation!

Additionally, all guides have been updated for clarification. Variants that are derivative of others now show enough of the rules in their own guide so that they don’t need to go to the original. For example, the Capahouse guide will also describe the new pieces rather than simply referring to the Capablanca guide.

*Upcoming additions*

As you've already seen, we've been adding arena tournaments regularly. We've already done our first shield tournament with Crazyhouse960, we'll have our first shield tournament with Shinobi on 9/2, and the second Crazyhouse960 after. With so many variants, we obviously will not be able to add tournaments for *everything*, but you can help by spreading the word to your friends and chess buddies. If you have specific requests for tournaments please let us know.

We'll also have a new addition to our lobby page very soon. Thanks a lot to VisualDennis/Tasshaq for implementing it!

That's all for now. For all those going back to school, good luck, but always feel free to spend your break time in Pychess!